# Faster equivalent value sorting