--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,2 @@
 example
-run_tests
+benchmark