--- a/benchmark.c
+++ b/benchmark.c
@@ -6,10 +6,24 @@
 #include "sort_r.h"
 
 
-bool printing = false;
+typedef enum {
+  SORT_METH_SIMPLE = 0,
+  SORT_METH_DEFAULT = 1,
+  SORT_METH_QSORT = 2
+} sort_type_t;
+
+bool printing = true;
+sort_type_t sort_type = SORT_METH_SIMPLE;
 int *randarr = NULL;
 
 static unsigned long cmp_counter = 0;
+
+static inline int qsort_cmp_ints(const void *aa, const void *bb)
+{
+  cmp_counter++;
+  const int a = *(const int*)aa, b = *(const int*)bb;
+  return (a > b ? 1 : -(a<b));
+}
 
 static inline int cmp_ints(const void *aa, const void *bb, void *arg)
 {
@@ -53,9 +67,21 @@ double run_time_trial(const char *name, int *arr, int n) {
   struct timespec start, finish;
   cmp_counter = 0;
   clock_gettime(CLOCK_REALTIME, &start);
-  sort_r_simple(arr, n, sizeof(int), cmp_ints, NULL);
+  switch(sort_type) {
+    case SORT_METH_SIMPLE:
+      sort_r_simple(arr, n, sizeof(int), cmp_ints, NULL);
+      break;
+    case SORT_METH_DEFAULT:
+      sort_r(arr, n, sizeof(int), cmp_ints, NULL);
+      break;
+    case SORT_METH_QSORT:
+      qsort(arr, n, sizeof(int), qsort_cmp_ints);
+      break;
+  }
   clock_gettime(CLOCK_REALTIME, &finish);
-  check_sorted(arr, n, sizeof(int), cmp_ints, NULL);
+  if(sort_type != SORT_METH_QSORT) {
+    check_sorted(arr, n, sizeof(int), cmp_ints, NULL);
+  }
 
   double timing_sec = get_timing(start, finish);
   report(name, n, timing_sec, cmp_counter);
@@ -83,11 +109,17 @@ double time_sorting_rand_vals(int *arr, int n) {
 }
 
 double time_sorting_mixed_vals(int *arr, int n) {
-  // Set a quarter of values to the same random value
+  /* Set a quarter of values to the same random value */
   for(int i = 0; i < n; i++) {
-    arr[i] = randarr[i] < n/4 ? randarr[n/2] : randarr[i];
+    arr[i] = (randarr[i] & 0x3) == 0 ? randarr[n/2] : randarr[i];
   }
   return run_time_trial("Mixed", arr, n);
+}
+
+double time_sorting_duped_vals(int *arr, int n) {
+  /* Repeat the same seq of random numbers e.g. [5,1,2,5,1,2]  */
+  for(int i = 0; i < n; i++) { arr[i] = randarr[i % (n/2)]; }
+  return run_time_trial("Duped", arr, n);
 }
 
 double run_tests(int *arr, int n) {
@@ -95,7 +127,8 @@ double run_tests(int *arr, int n) {
          time_sorting_ascending_vals(arr, n) +
          time_sorting_descending_vals(arr, n) +
          time_sorting_rand_vals(arr, n) +
-         time_sorting_mixed_vals(arr, n);
+         time_sorting_mixed_vals(arr, n) +
+         time_sorting_duped_vals(arr, n);
 }
 
 
@@ -104,22 +137,29 @@ int main()
   int n = 100000000;
   int *arr = malloc(n*sizeof(n));
 
+  printf("  %10s - %s\n", "Equal", "All values set to 10");
+  printf("  %10s - %s\n", "Ascending", "0..N-1");
+  printf("  %10s - %s\n", "Descending", "N-1..0");
+  printf("  %10s - %s\n", "Random", "N random numbers from rand()");
+  printf("  %10s - %s\n", "Mixed", "Random with 1/4 values set to the same value");
+  printf("  %10s - %s\n", "Duped", "Random array repeated e.g. [5,1,2,5,1,2]");
+
   srand(time(NULL));
   randarr = malloc(n*sizeof(n));
   for(int i = 0; i < n; i++) { randarr[i] = rand(); }
 
-  double run_secs = 0;
-  printing = false;
-
-  for(int i = 0; i < 20; i++) {
-    run_secs += run_tests(arr, i);
-  }
-
   /* Run tests using the entire array */
-  printing = true;
-  run_secs += run_tests(arr, n);
+  sort_type = SORT_METH_SIMPLE;
+  printf("\n == Using sort_r_simple()\n");
+  printf("Ran in %.2f s\n", run_tests(arr, n));
 
-  printf("Ran in %.2f s\n", run_secs);
+  printf("\n == Using sort_r()\n");
+  sort_type = SORT_METH_DEFAULT;
+  printf("Ran in %.2f s\n", run_tests(arr, n));
+
+  printf("\n == Using qsort()\n");
+  sort_type = SORT_METH_QSORT;
+  printf("Ran in %.2f s\n", run_tests(arr, n));
 
   free(arr);
   free(randarr);