Benchmarking this PR:

    ./benchmark
           Equal - All values set to 10
       Ascending - 0..N-1
      Descending - N-1..0
          Random - N random numbers from rand()
           Mixed - Random with 1/4 values set to the same value
           Duped - Random array repeated e.g. [5,1,2,5,1,2]

     == Using sort_r_simple()
           Equal: n=   100000000,    0.16 s,    200000000 comparisons (2.0X vs 26.6X)
       Ascending: n=   100000000,    4.06 s,   2583222807 comparisons (25.8X vs 26.6X)
      Descending: n=   100000000,    4.31 s,   2622161540 comparisons (26.2X vs 26.6X)
          Random: n=   100000000,   13.82 s,   3118756896 comparisons (31.2X vs 26.6X)
           Mixed: n=   100000000,   10.62 s,   2395846359 comparisons (24.0X vs 26.6X)
           Duped: n=   100000000,   13.19 s,   2995815278 comparisons (30.0X vs 26.6X)
    Ran in 46.16 s

     == Using sort_r()
           Equal: n=   100000000,    0.43 s,    200000010 comparisons (2.0X vs 26.6X)
       Ascending: n=   100000000,    0.78 s,    300000006 comparisons (3.0X vs 26.6X)
      Descending: n=   100000000,    3.86 s,   1100001339 comparisons (11.0X vs 26.6X)
          Random: n=   100000000,   21.29 s,   2912493212 comparisons (29.1X vs 26.6X)
           Mixed: n=   100000000,   16.26 s,   2225834496 comparisons (22.3X vs 26.6X)
           Duped: n=   100000000,   20.85 s,   2849404375 comparisons (28.5X vs 26.6X)
    Ran in 63.48 s

     == Using qsort()
           Equal: n=   100000000,    0.32 s,    100000011 comparisons (1.0X vs 26.6X)
       Ascending: n=   100000000,    0.58 s,    200000007 comparisons (2.0X vs 26.6X)
      Descending: n=   100000000,    2.91 s,   1000001340 comparisons (10.0X vs 26.6X)
          Random: n=   100000000,   18.73 s,   2812493213 comparisons (28.1X vs 26.6X)
           Mixed: n=   100000000,   14.33 s,   2125834497 comparisons (21.3X vs 26.6X)
           Duped: n=   100000000,   18.11 s,   2749404376 comparisons (27.5X vs 26.6X)
    Ran in 54.98 s


Before this PR changed sort_r_simple():

     == Using sort_r_simple()
           Equal: n=   100000000,   17.12 s,   2566445592 comparisons (25.7X vs 26.6X)
       Ascending: n=   100000000,    1.33 s,   2566445592 comparisons (25.7X vs 26.6X)
      Descending: n=   100000000,    2.20 s,   2596113945 comparisons (26.0X vs 26.6X)
          Random: n=   100000000,   18.42 s,   3065160214 comparisons (30.7X vs 26.6X)
           Mixed: n=   100000000,   18.29 s,   2943764556 comparisons (29.4X vs 26.6X)
           Duped: n=   100000000,   18.35 s,   3064155702 comparisons (30.6X vs 26.6X)
    Ran in 75.71 s